small fixes in Wiimote.framework

+++ trunk/Wiimote/WiimoteClassicController.h
@@ -15,7 +15,7 @@
     @private
         BOOL                                            m_ButtonState[WiimoteClassicControllerButtonCount];
         NSPoint                                         m_StickPositions[WiimoteClassicControllerStickCount];
-        float                                           m_AnalogShiftPositions[WiimoteClassicControllerAnalogShiftCount];
+        CGFloat                                         m_AnalogShiftPositions[WiimoteClassicControllerAnalogShiftCount];
         BOOL                                            m_IsCalibrationDataReaded;
         WiimoteDeviceClassicControllerCalibrationData   m_CalibrationData;
 }
